requirements.txt
gunicorn==20.0.4
Flask==1.1.2
firebase_admin==4.0.1
pylint==2.4.4
google-cloud-bigquery==1.27.2
google-cloud-firestore==1.9.0
Flask-Cors==3.0.10
google-cloud-storage==1.42.3
python-dotenv==0.19.2
matplotlib
pandas